o Use relative include directory for incs directive o Cross_compile is a nested directive

+++ tools/utils.py
@@ -4,56 +4,58 @@ import os
 
 """ This code is not in use """
 
-def build(objdir, resobj, flags, vars):
-
+def machine(vars, objdir):
 	machine = vars.get('machine')
 	if not machine:
 		print('machine is not set')
 		sys.exit(6)
 
-	src = os.path.join('..', 'mdepx/kernel', machine, 'include')
-	#if not os.path.exists(src):
-	#	print("machine headers not found")
-	#	sys.exit(7)
+	m = os.path.abspath(machine)
+	if not os.path.exists(m):
+		print("machine headers not found at path: %s" % m)
+		sys.exit(7)
 
 	dst = os.path.join(objdir, 'machine')
 	if os.path.exists(dst):
 		os.unlink(dst)
-	os.symlink(src, dst)
+	os.symlink(m, dst)
+
+def build(objdir, resobj, flags, vars):
+
+	machine(vars, objdir)
 
 	for obj in resobj:
 		fl = ' '.join(resobj[obj].get('cflags'))
 		for key in flags:
-			v = flags[key]
-			if v:
-				fl += ' -D%s=%s' % (key.upper(), v.upper())
-			else:
-				fl += ' -D%s' % (key.upper())
+			fl += ' -D%s' % (key.upper())
+			if flags[key]:
+				fl += '=%s' % flags[key].upper()
 
-		incs = '-I%s -I. -Imdepx/include -Imdepx/kernel -Imdepx/lib' %\
-			(objdir)
+		incs = '-I%s' % objdir
 		if 'incs' in resobj[obj]:
 			for inc in resobj[obj]['incs']:
-				incs += ' -Imdepx/%s ' % inc
+				incs += ' -I%s ' % inc
 
 		o = obj.replace(".o", ".c")
 		if not os.path.exists(o):
 			o = obj.replace(".o", ".S")
 
-		compiler = '%sgcc' % vars.get('cross_compile')
+		cc = resobj[obj].get('cross_compile') or ['']
+		compiler = '%sgcc' % cc[0]
 		objfile = os.path.join(objdir, obj)
 		os.makedirs(os.path.dirname(objfile), exist_ok=True)
 		cmd = '%s %s %s %s -c -o %s' % \
 			(compiler, fl, incs, o, objfile)
 
 		pcmd = "  CC      %s" % o
-		print(pcmd)
+		print(cmd)
 		status, output = getstatusoutput(cmd)
 		if status != 0:
 			print(output)
 			sys.exit(7)
 
-	linker = '%sld' % vars.get('cross_compile')
+	linker = resobj[obj].get('cross_compile') or ['']
+	linker = '%sld' % linker[0]
 	objs = ''
 	for o in resobj:
 		objs += "%s " % os.path.join(objdir, o)
